requirements.txt
opencv-python~=4.5.1
tensorflow~=2.5.0
tensorflow-gpu~=2.2.1
numpy~=1.19.5
matplotlib~=3.3.2
Keras~=2.4.3
Pillow~=8.2.0
streamlit~=0.82.0